(NPUP-15) Port assignment operator to the dispatcher.
Porting the assignment operator's implementation to the dispatcher.
