+++ lib/include/puppet/compiler/evaluation/operators/binary/assignment.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file
+ * Declares the assignment operator.
+ */
+#pragma once
+
+#include "descriptor.hpp"
+
+namespace puppet { namespace compiler { namespace evaluation { namespace operators { namespace binary {
+
+    /**
+     * Implements the assignment operator.
+     */
+    struct assignment
+    {
+        /**
+         * Create a binary operator descriptor.
+         * @return Returns the binary operator descriptor representing this operator.
+         */
+        static descriptor create_descriptor();
+    };
+
+}}}}}  // namespace puppet::compiler::evaluation::operators::binary
